net-tcp_bbr: v2: additive component for non-CE-marked bw probing
Fairness with pacing-based ECN response ecn_reprobe_swiftly (earlier
in this test branch) is probably insufficient.

This experimental patch (with an ecn_reprobe_additive knob to
enable/disable) gives an additive increase to both bw_hi and
inflight_hi when starting an ecn_reprobe_swiftly probe.  This serves
as a sort of "universal basic income" that provides an additive
component to the dynamics.

Why add this? Chiu and Jain [1] says that a MIMD-like scheme (like
BBRv2) can work as long as there is an additive component as well.
Specifically, on page 9 it says:

"Proposition 1. In order to satisfy the requirements of distributed
convergence to efficiency and fairness without truncation, the linear
decrease policy should be multiplicative, and the linear increase
policy should always have an additive component, and optionally it may
have a multiplicative component with the coefficient no less than
one."

So basically, according to [1] you need MD, and AI, but you can also
have an optional MI component as well.

[1] D.-M. Chiu und R. Jain, #Analysis of the increase and decrease
algorithms for congestion avoidance in computer networks", Computer
Networks and ISDN Systems, Bd. 17, Nr. 1, S. 1-14, Juni 1989.

Change-Id: If853cc4507457e2cb4c771a1e56db0249f3321df
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>